+++ apps/server/src/modules/board/service/card.service.spec.ts
@@ -1,6 +1,7 @@
 import { createMock, DeepMocked } from '@golevelup/ts-jest';
 import { NotFoundException } from '@nestjs/common';
 import { Test, TestingModule } from '@nestjs/testing';
+import { TextElement } from '@shared/domain';
 import { setupEntities } from '@shared/testing';
 import {
 	cardFactory,
@@ -100,55 +101,11 @@ describe(CardService.name, () => {
 		});
 
 		it('should throw an error if some DOs are not cards', async () => {
-			const textElements = textElementFactory.buildList(2);
+			const textElements = textElementFactory.buildList(2) as TextElement[];
 			const textElementIds = textElements.map((t) => t.id);
 			boardDoRepo.findByIds.mockResolvedValue(textElements);
 
 			await expect(service.findByIds(textElementIds)).rejects.toThrow();
-		});
-	});
-
-	describe('creating a card', () => {
-		const setup = () => {
-			const column = columnFactory.build();
-			const board = columnBoardFactory.build({ children: [column] });
-			const boardId = board.id;
-			const columnId = column.id;
-
-			return { board, boardId, column, columnId };
-		};
-
-		it('should save a list of cards using the repo', async () => {
-			const { board, boardId, columnId } = setup();
-
-			boardDoRepo.findById.mockResolvedValueOnce(board);
-
-			await service.createCard(boardId, columnId);
-
-			expect(boardDoRepo.save).toHaveBeenCalledWith(
-				[
-					expect.objectContaining({
-						id: expect.any(String),
-						title: '',
-						height: 150,
-						children: [],
-						createdAt: expect.any(Date),
-						updatedAt: expect.any(Date),
-					}),
-				],
-				columnId
-			);
-		});
-
-		it('should throw not found exception if requested column id has not been found', async () => {
-			const { board, boardId } = setup();
-
-			const notExistingColumnId = new ObjectId().toHexString();
-			const error = new NotFoundException(`The requested Column: id='${notExistingColumnId}' has not been found.`);
-
-			boardDoRepo.findById.mockResolvedValueOnce(board);
-
-			await expect(service.createCard(boardId, notExistingColumnId)).rejects.toThrowError(error);
 		});
 	});
 });
